Some types should be ulong, not int

+++ net/dll/seal/batchencoder.cpp
@@ -41,7 +41,7 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Destroy(void* thisptr)
     return S_OK;
 }
 
-SEALDLL HRESULT SEALCALL BatchEncoder_Encode1(void* thisptr, int count, uint64_t* values, void* destination)
+SEALDLL HRESULT SEALCALL BatchEncoder_Encode1(void* thisptr, uint64_t count, uint64_t* values, void* destination)
 {
     BatchEncoder* encoder = FromVoid<BatchEncoder>(thisptr);
     IfNullRet(encoder, E_POINTER);
@@ -66,7 +66,7 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Encode1(void* thisptr, int count, uint64_t
     }
 }
 
-SEALDLL HRESULT SEALCALL BatchEncoder_Encode2(void* thisptr, int count, int64_t* values, void* destination)
+SEALDLL HRESULT SEALCALL BatchEncoder_Encode2(void* thisptr, uint64_t count, int64_t* values, void* destination)
 {
     BatchEncoder* encoder = FromVoid<BatchEncoder>(thisptr);
     IfNullRet(encoder, E_POINTER);
@@ -112,7 +112,7 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Encode3(void* thisptr, void* plain, void* 
     }
 }
 
-SEALDLL HRESULT SEALCALL BatchEncoder_Decode1(void* thisptr, void* plain, int* count, uint64_t* destination, void* pool)
+SEALDLL HRESULT SEALCALL BatchEncoder_Decode1(void* thisptr, void* plain, uint64_t* count, uint64_t* destination, void* pool)
 {
     BatchEncoder* encoder = FromVoid<BatchEncoder>(thisptr);
     IfNullRet(encoder, E_POINTER);
@@ -128,7 +128,7 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Decode1(void* thisptr, void* plain, int* c
         vector<uint64_t> result;
         encoder->decode(*plainptr, result, *handle);
 
-        *count = static_cast<int>(result.size());
+        *count = result.size();
 
         if (nullptr == destination)
         {
@@ -149,7 +149,7 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Decode1(void* thisptr, void* plain, int* c
     }
 }
 
-SEALDLL HRESULT SEALCALL BatchEncoder_Decode2(void* thisptr, void* plain, int* count, int64_t* destination, void* pool)
+SEALDLL HRESULT SEALCALL BatchEncoder_Decode2(void* thisptr, void* plain, uint64_t* count, int64_t* destination, void* pool)
 {
     BatchEncoder* encoder = FromVoid<BatchEncoder>(thisptr);
     IfNullRet(encoder, E_POINTER);
@@ -165,7 +165,7 @@ SEALDLL HRESULT SEALCALL BatchEncoder_Decode2(void* thisptr, void* plain, int* c
         vector<int64_t> result;
         encoder->decode(*plainptr, result, *handle);
 
-        *count = static_cast<int>(result.size());
+        *count = result.size();
 
         if (nullptr == destination)
         {
